CURRENT_TIMESTAMP/CURRENT_DATE/CURRENT_TIME
