+++ query/src/org/labkey/query/sql/SqlBase.g
@@ -174,6 +174,9 @@ REGR_SXY : 'regr_sxy';
 REGR_SYY : 'regr_syy';
 COUNT : 'count';
 CROSS : 'cross';
+CURRENT_DATE : 'current_date';
+CURRENT_TIME : 'current_time';
+CURRENT_TIMESTAMP : 'current_timestamp';
 DELETE : 'delete';
 DISTINCT : 'distinct';
 DOT : '.';
@@ -755,6 +758,10 @@ starAtom
 primaryExpression
 	:   ARRAY exprList ']' -> ^(METHOD_CALL IDENT["ARRAY_CONSTRUCT"] exprList)
 	|   TEXTARRAY exprList ']' -> ^(METHOD_CALL IDENT["TEXTARRAY_CONSTRUCT"] exprList)
+	// SQL-standard niladic datetime keywords -- no parens allowed; sugar for curdate()/curtime()/now()
+	|   CURRENT_DATE -> ^(METHOD_CALL IDENT["CURDATE"] ^(EXPR_LIST))
+	|   CURRENT_TIME -> ^(METHOD_CALL IDENT["CURTIME"] ^(EXPR_LIST))
+	|   CURRENT_TIMESTAMP -> ^(METHOD_CALL IDENT["NOW"] ^(EXPR_LIST))
 	|   id=identPrimary
 	|   constant
 	|   OPEN! ( expression | subQuery) CLOSE!
